==> Estudio_crianzas_Chile/01_03_tasas.ipynb <==
# Se cargan las librerías y tabla de datos
import pandas as pd
import plotly.express as px
from plotly import graph_objects as go
import os
# os.getcwd()
os.chdir('/home/clautc/Proyectos_/Estudios/Estudio_crianzas_Chile')
from plotly.subplots import make_subplots
ruta = 'data/registros_estadisticos/df_edades.csv'
df = pd.read_csv(ruta)
df.drop(columns='Unnamed: 0', inplace=True)



pob = pd.read_csv('data/registros_estadisticos/proyecciones_1992_2035_data.csv', sep=';', thousands='.')

df['grupo_dec_mujer'] = ''
df.loc[df['grupo_etario_madre'] == 'MENORES 15 AÑO', 'grupo_dec_mujer'] = '19 o menos'
df.loc[df['grupo_etario_madre'] == '15 A 19 AÑOS', 'grupo_dec_mujer'] = '19 o menos'
df.loc[df['grupo_etario_madre'] == '20 A 24 AÑOS', 'grupo_dec_mujer'] = '20 a 29'
df.loc[df['grupo_etario_madre'] == '25 A 29 AÑOS', 'grupo_dec_mujer'] = '20 a 29'
df.loc[df['grupo_etario_madre'] == '30 A 34 AÑOS', 'grupo_dec_mujer'] = '30 a 39'
df.loc[df['grupo_etario_madre'] == '35 A 39 AÑOS', 'grupo_dec_mujer'] = '30 a 39'
df.loc[df['grupo_etario_madre'] == '40 A 44 AÑOS', 'grupo_dec_mujer'] = '40 a 49'
df.loc[df['grupo_etario_madre'] == '45 A 49 AÑOS', 'grupo_dec_mujer'] = '40 a 49'
df.loc[df['grupo_etario_madre'] == '50 O MAS AÑOS', 'grupo_dec_mujer'] = '50 o más'

df['grupo_quin_mujer'] = ''
df.loc[df['grupo_etario_madre'] == 'MENORES 15 AÑO', 'grupo_quin_mujer'] = '19 o menos'
df.loc[df['grupo_etario_madre'] == '15 A 19 AÑOS', 'grupo_quin_mujer'] = '19 o menos'
df.loc[df['grupo_etario_madre'] == '20 A 24 AÑOS', 'grupo_quin_mujer'] = '20 a 24'
df.loc[df['grupo_etario_madre'] == '25 A 29 AÑOS', 'grupo_quin_mujer'] = '25 a 29'
df.loc[df['grupo_etario_madre'] == '30 A 34 AÑOS', 'grupo_quin_mujer'] = '30 a 34'
df.loc[df['grupo_etario_madre'] == '35 A 39 AÑOS', 'grupo_quin_mujer'] = '35 a 39'
df.loc[df['grupo_etario_madre'] == '40 A 44 AÑOS', 'grupo_quin_mujer'] = '40 a 44'
df.loc[df['grupo_etario_madre'] == '45 A 49 AÑOS', 'grupo_quin_mujer'] = '45 a 49'
df.loc[df['grupo_etario_madre'] == '50 O MAS AÑOS', 'grupo_quin_mujer'] = '45 a 49'

df['grupo_etario_padre'] = df['grupo_etario_padre'].str.strip()
df['grupo_dec_hombre'] = 'NO ESPECIFICADO'
df.loc[df['grupo_etario_padre'] == 'MENORES 15 AÑOS', 'grupo_dec_hombre'] = '19 o menos'
df.loc[df['grupo_etario_padre'] == '15 A 19 AÑOS',     'grupo_dec_hombre'] = '19 o menos'
df.loc[df['grupo_etario_padre'] == '20 A 24 AÑOS',     'grupo_dec_hombre'] = '20 a 29'
df.loc[df['grupo_etario_padre'] == '25 A 29 AÑOS',     'grupo_dec_hombre'] = '20 a 29'
df.loc[df['grupo_etario_padre'] == '30 A 34 AÑOS',     'grupo_dec_hombre'] = '30 a 39'
df.loc[df['grupo_etario_padre'] == '35 A 39 AÑOS',     'grupo_dec_hombre'] = '30 a 39'
df.loc[df['grupo_etario_padre'] == '40 A 44 AÑOS',     'grupo_dec_hombre'] = '40 a 49'
df.loc[df['grupo_etario_padre'] == '45 A 49 AÑOS',     'grupo_dec_hombre'] = '40 a 49'
df.loc[df['grupo_etario_padre'] == '50 O MAS AÑOS',   'grupo_dec_hombre'] = '50 o más'

mask = df.grupo_etario_madre.isin(['NO ESPECIFICADO'])
df = df[~mask]  # se eliminan las edades no especificadas de mujeres

# pd.crosstab(df.grupo_etario_padre, df.grupo_dec_hombre)

df.loc[df['grupo_etario_madre'] == 'MENORES 15 AÑO', 'grupo_etario_madre'] = 'MENORES 15 AÑOS'
df['ano_nac'] = pd.to_datetime(df['ano_nac'], format='%Y')
df = df.sort_values('ano_nac').reset_index(drop=True)
df['anio'] = df.ano_nac.dt.year

 # Las tasas se sacan con la población de 15 a 49 años de edad
# Se define como el número promedio de hijos por cada mil mujeres según su grupo de edad
pob = pob.loc[pob.edad >14]
pob = pob.loc[pob.edad<50]
 # solo afecta a las mujeres
pobm = pob.loc[pob.sexo=='mujer']
pobm = pobm.reset_index(drop=True)
# crear variable grupo edad por diez

# pobm['grupo_dec'] = ''
# pobm.loc[pobm.edad  < 20, 'grupo_dec'] = '19 o menos'
# pobm.loc[(pobm.edad > 19) & (pobm.edad < 30), 'grupo_dec'] = '20 a 29'
# pobm.loc[(pobm.edad > 29) & (pobm.edad < 40), 'grupo_dec'] = '30 a 39'
# pobm.loc[(pobm.edad > 39) & (pobm.edad < 50), 'grupo_dec'] = '40 a 49'
# pobm.loc[pobm.edad  > 49, 'grupo_dec'] = '50 o más'

pobm['grupo_quin'] = ''
pobm.loc[pobm.edad  < 20, 'grupo_quin'] = '19 o menos'
pobm.loc[(pobm.edad > 19) & (pobm.edad < 25), 'grupo_quin'] = '20 a 24'
pobm.loc[(pobm.edad > 24) & (pobm.edad < 30), 'grupo_quin'] = '25 a 29'
pobm.loc[(pobm.edad > 29) & (pobm.edad < 35), 'grupo_quin'] = '30 a 34'
pobm.loc[(pobm.edad > 34) & (pobm.edad < 40), 'grupo_quin'] = '35 a 39'
pobm.loc[(pobm.edad > 39) & (pobm.edad < 45), 'grupo_quin'] = '40 a 44'
pobm.loc[(pobm.edad > 44) & (pobm.edad < 50), 'grupo_quin'] = '45 a 49'



pobm = pobm.drop(columns=['edad', 'sexo'])


poblacion_quin = pobm.melt(id_vars=['grupo_quin']).groupby(['grupo_quin', 'variable'])['value'].sum().reset_index()

poblacion_quin.columns = ['gr', 'variable', 'pop']
poblacion_quin['anio'] = pd.to_datetime(poblacion_quin['variable'], format='%Y')
poblacion_quin = poblacion_quin.sort_values('anio').reset_index(drop=True)
poblacion_quin['anio'] = poblacion_quin.anio.dt.year
poblacion_quin = poblacion_quin.loc[poblacion_quin.anio<2020]
# poblacion_quin

poblacion_tot= pobm.melt(id_vars=['grupo_quin']).groupby(['variable'])['value'].sum().reset_index()
poblacion_tot['anio'] = pd.to_datetime(poblacion_tot['variable'], format='%Y')
poblacion_tot = poblacion_tot.sort_values('anio').reset_index(drop=True)
poblacion_tot['anio'] = poblacion_tot.anio.dt.year

poblacion_quin


df.value_counts(subset=['anio']).reset_index(name='nacimientos')
tef = pd.merge(df.value_counts(subset=['anio', 'grupo_quin_mujer']).reset_index(name='nacimientos'),
               poblacion_quin.drop(columns='variable'),
               left_on=['anio', 'grupo_quin_mujer'], right_on=['anio', 'gr'])
tef['tef'] = tef.nacimientos/tef['pop']
tef['tef_mil'] = tef.nacimientos/tef['pop'] *1000
gr = tef.groupby('anio')
tef['tgf'] = gr['tef'].transform(sum)*5
# len(tef.grupo_quin_mujer.unique())
df_tef = tef.sort_values('anio').reset_index(drop=True)
df_tef

df_tef['orden_edad'] = 0
df_tef.loc[df_tef.gr.str.match('^[1]9'), 'orden_edad'] = 1
df_tef.loc[df_tef.gr.str.match('^[2]0'), 'orden_edad'] = 2
df_tef.loc[df_tef.gr.str.match('^[2]5'), 'orden_edad'] = 3
df_tef.loc[df_tef.gr.str.match('^[3]0'), 'orden_edad'] = 4
df_tef.loc[df_tef.gr.str.match('^[3]5'), 'orden_edad'] = 5
df_tef.loc[df_tef.gr.str.match('^[4]0'), 'orden_edad'] = 6
df_tef.loc[df_tef.gr.str.match('^[4]5'), 'orden_edad'] = 7
df_tef = df_tef.sort_values(['anio', 'orden_edad']).reset_index(drop=True)

# df_tef
# anios = df_tef.anio.unique()
anios = [1992,  1999, 2009,  2019]
grupos = df_tef.gr.unique()
shift = [45, 15, -25, -45]
posiciones = ['top left', 'middle right', 'bottom left', 'bottom right']
colores = ['#1F77B4', '#FF7F0E', '#2CA02C', '#D62728']

La tasa específica de fecundidad refiere al número de nacimientos que ocurren durante un determinado año o periodo de
referencia por cada mil mujeres en edad reproductiva (15 a 49 años de edad) clasificada en grupos de edad simple o
quinquenales.

Esta tasa sirve como medida de la frecuencia relativa de las personas nacidas vivas tenidas por mujeres de diferentes
 edades durante sus años reproductivos. También es un cómputo intermedio para calcular la tasa global de fecundidad.
 La tasa global de fecundidad es el número de personas que nacerían por mujer o por mil mujeres, si ella o ellas
 tuvieran sus hijos o hijas en todos sus años reproductivos.

Es un indicador que no se ve afectado por la composición de la población por edad y sexo y proporciona una medida
fácil de comprender de la fecundidad hipotética completa.





fig = go.Figure()

for a in range(len(anios)):
    y = df_tef.loc[df_tef['anio']==anios[a], 'tef_mil']
    x = df_tef.loc[df_tef['anio']==anios[a], 'gr']

    print(y)
    print(x)
    orden_ = df_tef.loc[df_tef['anio']==anios[a], 'orden_edad'].unique()[0]

    fig.add_trace(go.Scatter(
        x = x,
        y = y,
        name = str(anios[a]),
        # marker=dict(color=color),
        mode='lines+markers',
        legendrank=orden_
    )
    )

    for g in grupos:
        fig.add_annotation(
            x = df_tef.loc[(df_tef['anio']==anios[a]) & (df_tef.gr==g), 'gr'].to_list()[0],
            y = df_tef.loc[(df_tef['anio']==anios[a]) & (df_tef.gr==g), 'tef_mil'].to_list()[0],
            yshift = shift[a],
            text = str(round(df_tef.loc[(df_tef['anio']==anios[a]) & (df_tef.gr==g), 'tef_mil'].to_list()[0])),
            showarrow=False,
            bordercolor=colores[a],
            borderwidth=2,
            borderpad=4,
            bgcolor=colores[a],
            font= dict(
                size=16,
                color='white',

            )
    )
    # for i in df_tef.gr.to_list():
    #     fig.add_annotation( x=i,
    #                         y=y.to_list()[0],
    #                         text = str(round(y.to_list()[0], 1)),
    #                         font= dict(
    #                             size=16,
    #                             color='white',
    #                         ),
    #                         showarrow=False,
    #                         # yshift=20,
    #                         bordercolor=colores[a],
    #                         borderwidth=2,
    #                         borderpad=4,
    #                         bgcolor=colores[a]
    #                         )


fig.update_layout(template='none')
fig.update_yaxes(title='Tasas Específicas de Fecundidad')
fig.update_xaxes(title='Grupos de edades quinquenales', dtick='M12',  ticklabelmode='period')
fig.write_image('resultados/img/02_tef.jpeg')
fig.show()

df_global = df_tef[['anio', 'tgf']]
df_global = df_global.drop_duplicates().reset_index(drop=True)
an_x1 = df_global.loc[df_global.anio==1992, 'anio']
an_y1 = df_global.loc[df_global.anio==1992, 'tgf']
an_x2 = df_global.loc[df_global.anio==1999, 'anio']
an_y2 = df_global.loc[df_global.anio==1999, 'tgf']
an_x3 = df_global.loc[df_global.anio==2009, 'anio']
an_y3 = df_global.loc[df_global.anio==2009, 'tgf']
an_x4 = df_global.loc[df_global.anio==2019, 'anio']
an_y4 = df_global.loc[df_global.anio==2019, 'tgf']
print(an_y2.to_list()[0])

fig = go.Figure()
fig.add_trace(
    go.Scatter(
        x = df_global.anio,
        y = df_global.tgf,
        mode='lines+markers',
    )
)
fig.update_layout(template='none')
fig.update_yaxes(title='Tasa Global de Fecundidad')
fig.update_xaxes(title='Periodo de estudio', dtick='M12',  ticklabelmode='period')
fig.add_annotation(x=1992, y=an_y1[0],
                   # yref= 'paper',
                   # xref='paper',
                   text = str(round(an_y1[0], 1)),
                   font= dict(
                    size=16,
                       color='#E622DD'
                   ),
                   showarrow=False,
                   yshift=20,
                   bordercolor='#E622DD',
                   borderwidth=2,
                   borderpad=4,
                   bgcolor='#A3E4D7'
                   )
fig.add_annotation(x=1999, y=an_y2.to_list()[0],
                  # yref= 'paper',
                  # xref='paper',
                  text = str(round(an_y2.to_list()[0], 1)),
                  font= dict(
                      size=16,
                      color='#E622DD'
                  ),
                  showarrow=False,
                  yshift=20,
                  bordercolor='#E622DD',
                  borderwidth=2,
                  borderpad=4,
                  bgcolor='#A3E4D7'
                  )

fig.add_annotation(x=2009, y=an_y3.to_list()[0],
                   # yref= 'paper',
                   # xref='paper',
                   text = str(round(an_y3.to_list()[0], 1)),
                   font= dict(
                       size=16,
                       color='#E622DD'
                   ),
                   showarrow=False,
                   yshift=20,
                   bordercolor='#E622DD',
                   borderwidth=2,
                   borderpad=4,
                   bgcolor='#A3E4D7'
                   )
fig.add_annotation(x=2019, y=an_y4.to_list()[0],
                   # yref= 'paper',
                   # xref='paper',
                   text = str(round(an_y4.to_list()[0], 1)),
                   font= dict(
                       size=16,
                       color='#E622DD'
                   ),
                   showarrow=False,
                   yshift=20,
                   bordercolor='#E622DD',
                   borderwidth=2,
                   borderpad=4,
                   bgcolor='#A3E4D7'
                   )
fig.write_image('resultados/img/01_tgf.jpeg')
fig.show()
