==> Estudio_crianzas_Chile/01_04_estadocivil.ipynb <==
# Se cargan las librerías y tabla de datos
import pandas as pd
import plotly.express as px
from plotly import graph_objects as go
import os
# os.getcwd()
os.chdir('/home/clautc/Proyectos_/Estudios/Estudio_crianzas_Chile')
from plotly.subplots import make_subplots
ruta = 'data/registros_estadisticos/df_geograf.csv'
df = pd.read_csv(ruta)

df.groupby(['ano_nac', 'est_civ_madre'])['grupo_etario_madre'].count()

estado_civil_madre= {
    2: 'Casada',
    1: 'Soltera',
    3: 'Otros (viuda, ignorada)',
    9: 'Otros (viuda, ignorada)'
}
estado_civil_color= {
    1: '#9A208C',
    2: '#E11299',
    3: '#00235B',
    9: '#E21818'
}
df['ecm'] = df['est_civ_madre'].map(lambda x: estado_civil_madre.get(x))
df['anio'] = pd.to_datetime(df.ano_nac, format='%Y')
df['anio'] = df.anio.dt.year


dfn = df.value_counts(subset=['anio', 'ecm']).reset_index(name='conteo').sort_values('anio').reset_index(drop=True)
dfn

fig = go.Figure()
for c in dfn.ecm.unique().tolist():
    fig.add_trace(
        go.Scatter(
            x = dfn.loc[dfn.ecm == c, 'anio'],
            y = dfn.loc[dfn.ecm == c, 'conteo'],
            mode='lines',
            name= c
        )
    )
fig.update_layout(template='none')
fig.update_yaxes(title='Número de nacimientos')
fig.update_xaxes(title='Periodo de estudio', dtick='M12',
                 # ticklabelmode='period'
                 )
fig.write_image('resultados/img/01_estadocivilmadre.jpeg')
fig.show()
