==> Estudio_crianzas_Chile/02_quintilies_hogares_con_menores.ipynb <==
# Se cargan las librerías y tabla de datos
import pandas as pd
import plotly.express as px
from plotly import graph_objects as go
import os
# os.getcwd()
os.chdir('/home/clautc/Proyectos_/Estudios'
         '/Estudio_crianzas_Chile')
df = pd.read_csv('resultados/tablas/02_quintil_menor_18.csv')
sinmenores = 'sin menores'
conmenores = 'con menores'
tipoest = 'conteo'

df_sinmenores = df.query('menor_18 == @sinmenores and tipo_est == @tipoest')
df_conmenores = df.query('menor_18 == @conmenores and tipo_est == @tipoest')
smgrupoanio = df_sinmenores.groupby('anio')
df_sinmenores['tot'] = smgrupoanio['dato'].transform('sum')
df_sinmenores['prop'] = df_sinmenores.dato/df_sinmenores.tot*100
cmgrupoanio = df_conmenores.groupby('anio')
df_conmenores['tot'] = cmgrupoanio['dato'].transform('sum')
df_conmenores['prop'] = df_conmenores.dato/df_conmenores.tot*100


# anios= df.anio.unique()
anios = [1990, 1994, 1998, 2003, 2009, 2015, 2020]
shift = [-25, 20, 20, 20, 20]
posiciones = ['bottom center', 'middle center', 'middle center', 'middle center', 'bottom left']
colores = ['#1F77B4', '#FF7F0E', '#2CA02C', '#D62728', '#9467BD']
fig = go.Figure()
quintiles = df_sinmenores.quintil.unique()
for q in range(len(quintiles)):
    fig.add_trace(
        go.Scatter(
             x = df_sinmenores.loc[df_sinmenores.quintil == quintiles[q], 'anio'],
             y = df_sinmenores.loc[df_sinmenores.quintil == quintiles[q], 'prop'],
             mode='lines + markers',
            name=quintiles[q]
         )
     )
    for a in anios:
        fig.add_annotation(
                x = df_sinmenores.loc[(df_sinmenores.quintil == quintiles[q]) & (df_sinmenores.anio == a), 'anio'].to_list()[0],
                y = df_sinmenores.loc[(df_sinmenores.quintil == quintiles[q]) & (df_sinmenores.anio == a), 'prop']
                .to_list()[0],
                yshift = shift[q],
                text = str(a) + ':' + '\n' + str(round(df_sinmenores.loc[(df_sinmenores.quintil == quintiles[q]) &
                                                                 (df_sinmenores
                                                                                                    .anio ==
                                                                                              a), 'prop'].to_list()
                                 [0])),
                showarrow=False,
                bordercolor=colores[q],
                borderwidth=2,
                borderpad=4,
                bgcolor=colores[q],
                font= dict(
                    size=10,
                    color='white',

                )
        )
fig.update_layout(template='none')
fig.write_image('resultados/img/02_sinmenoresquintil.jpeg')

fig.show()

# anios= df.anio.unique()
anios = [1990, 1994, 1998, 2003, 2009, 2015, 2020]
shift = [20, -20, 0, 0, 0]
posiciones = ['top right', 'bottom left', 'middle center', 'middle center', 'middle left']
colores = ['#1F77B4', '#FF7F0E', '#2CA02C', '#D62728', '#9467BD']
fig = go.Figure()
quintiles = df_conmenores.quintil.unique()
for q in range(len(quintiles)):
    fig.add_trace(
        go.Scatter(
            x = df_conmenores.loc[df_conmenores.quintil == quintiles[q], 'anio'],
            y = df_conmenores.loc[df_conmenores.quintil == quintiles[q], 'prop'],
            mode='lines + markers',
            name=quintiles[q]
        )
    )
    for a in anios:
        fig.add_annotation(
            x = df_conmenores.loc[(df_conmenores.quintil == quintiles[q]) & (df_conmenores.anio == a), 'anio'].to_list
            ()[0],
            y = df_conmenores.loc[(df_conmenores.quintil == quintiles[q]) & (df_conmenores.anio == a), 'prop']
            .to_list()[0],
            yshift = shift[q],
            text = str(a) + ':' + '\n' + str(round(df_conmenores.loc[(df_conmenores.quintil == quintiles[q]) &
                                                                     (df_conmenores
                                                                      .anio ==
                                                                      a), 'prop'].to_list()
                                                   [0])),
            showarrow=False,
            bordercolor=colores[q],
            borderwidth=2,
            borderpad=4,
            bgcolor=colores[q],
            font= dict(
                size=10,
                color='white',

            )
        )
fig.update_layout(template='none')
fig.write_image('resultados/img/02_conmenoresquintil.jpeg')

fig.show()
